fix: ensure all useful types are exported
