--- src/serializers/index.ts
@@ -1,2 +1,0 @@
-// * The transitive dependency "pretty-format" is a dependency of Jest
-export type { Plugin as SnapshotSerializer } from 'pretty-format';
